pyproject.toml
[project]
name = "proofsanalyze"
version = "0.2.0"
description = "Async FastAPI service for mathematical proof verification with LLM agents and RAG"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "google-genai>=1.0.0,<2.0.0",
    "protobuf>=4.25.8,<6.0.0",
    "chromadb>=0.4.24,<1.0.0",
    "sentence-transformers>=3.0.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<1.0.0",
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.6.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["static/*", "core/clients/db/knowledge_base/*.json"]
